requirements.txt
berserk
python-chess
matplotlib
seaborn
pandas